[DEV] CRUD (tanpa delete)

+++ app/Http/Controllers/Admin/PolicyKesehatanController.php
@@ -20,6 +20,9 @@ use Symfony\Component\HttpFoundation\Response;
 use Yajra\DataTables\Facades\DataTables;
 use Illuminate\Support\Facades\DB;
 
+use App\Http\Requests\StorePolicyUnifiedRequest;
+use App\Http\Requests\UpdatePolicyUnifiedRequest;
+
 class PolicyKesehatanController extends Controller
 {
     use MediaUploadingTrait, CsvImportTrait;
@@ -135,32 +138,14 @@ class PolicyKesehatanController extends Controller
         return view('admin.policyKesehatans.create', compact('assigned_to_customers', 'assigned_to_users', 'insurance_products', 'isAdmin'));
     }
 
-    public function store(Request $request)
+    public function store(StorePolicyUnifiedRequest $request, PolicyKesehatan $policyKesehatan)
     {
-
-        DB::beginTransaction();
-
+        abort_if(Gate::denies('policy_kesehatan_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        
         try {
-            $polisCentral = PoliciesCentral::create($request->only(
-                                            [
-                                                'assigned_to_customer_id',
-                                                'policy_number',
-                                                'policy_number_external',
-                                                'insurance_product_id',
-                                                'start_date',
-                                                'end_date',
-                                                'premium_amount',
-                                                'discount',
-                                                'discount_total',
-                                                'aksessoris_tambahan',
-                                                'aksesoris_harga',
-                                                'biaya_lainnya',
-                                                'sum_insured',
-                                                'policy_status',
-                                                'payment_status',
-                                                'assigned_to_user_id',
-                                            ])
-                            );
+            DB::beginTransaction();
+            // Insert ke tabel central
+            $policiesCentral = PoliciesCentral::create($request->centralData());
 
             foreach ($request->input('external_polis_doc', []) as $file) {
                 $policiesCentral->addMedia(storage_path('tmp/uploads/' . basename($file)))->toMediaCollection('external_polis_doc');
@@ -170,30 +155,16 @@ class PolicyKesehatanController extends Controller
                 Media::whereIn('id', $media)->update(['model_id' => $policiesCentral->id]);
             }
 
-            $policyRumahGedung = PolicyRumahGedung::create(['id_policies_id' => $polisCentral->id] + $request->only(
-                [
-                    'insurance_product_id',
-                    'lokasi_pertanggungan',
-                    'jenis_rumah_gedung_id',
-                    'keterangan',
-                    'jenis_paket_id',
-                    'nama_tertanggung',
-                    'ttl_tertanggung',
-                    'alamat_tertanggung',
-                    'email',
-                    'no_phone',
-                    'nama_paket',
-                    'assigned_to_user_id',
-                    'assigned_to_customer_id',
-                ]
-            ));
+            $childData = $request->childData();
+            $childData['id_policies_id'] = $policiesCentral->id;
+            $policyKesehatan->create($childData);
 
             foreach ($request->input('upload_dokumen', []) as $file) {
-                $policyRumahGedung->addMedia(storage_path('tmp/uploads/' . basename($file)))->toMediaCollection('upload_dokumen');
+                $policyKesehatan->addMedia(storage_path('tmp/uploads/' . basename($file)))->toMediaCollection('upload_dokumen');
             }
 
             if ($media = $request->input('ck-media', false)) {
-                Media::whereIn('id', $media)->update(['model_id' => $policyRumahGedung->id]);
+                Media::whereIn('id', $media)->update(['model_id' => $policyKesehatan->id]);
             }
 
             DB::commit(); // simpan semua perubahan
@@ -211,38 +182,77 @@ class PolicyKesehatanController extends Controller
     {
         abort_if(Gate::denies('policy_kesehatan_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        $id_policies = PoliciesCentral::pluck('policy_number', 'id')->prepend(trans('global.pleaseSelect'), '');
+        $assigned_to_customers = CrmCustomer::pluck('first_name', 'id')->prepend(trans('global.pleaseSelect'), '');
 
         $insurance_products = InsuranceProduct::pluck('product_name', 'id')->prepend(trans('global.pleaseSelect'), '');
 
-        $assigned_to_users = User::pluck('name', 'id')->prepend(trans('global.pleaseSelect'), '');
+        $assigned_to_users = User::where('id', '!=', 1)->pluck('name', 'id')->prepend(trans('global.pleaseSelect'), '');
 
-        $assigned_to_customers = CrmCustomer::pluck('first_name', 'id')->prepend(trans('global.pleaseSelect'), '');
+        $isAdmin = auth()->user()->roles->contains(1);
 
-        $policyKesehatan->load('id_policies', 'insurance_product', 'assigned_to_user', 'assigned_to_customer', 'created_by');
+        //$policyKesehatan->load('id_policies', 'insurance_product', 'assigned_to_user', 'assigned_to_customer', 'created_by');
 
-        return view('admin.policyKesehatans.edit', compact('assigned_to_customers', 'assigned_to_users', 'id_policies', 'insurance_products', 'policyKesehatan'));
+        return view('admin.policyKesehatans.edit', [
+            'policiesCentral' => $policyKesehatan->id_policies,
+            'assigned_to_customers' => $assigned_to_customers,
+            'insurance_products' => $insurance_products,
+            'assigned_to_users' => $assigned_to_users,
+            'policyKesehatan' => $policyKesehatan,
+            'isAdmin' => $isAdmin]);
     }
 
-    public function update(UpdatePolicyKesehatanRequest $request, PolicyKesehatan $policyKesehatan)
+    public function update(UpdatePolicyUnifiedRequest $request, PolicyKesehatan $policyKesehatan)
     {
-        $policyKesehatan->update($request->all());
+        abort_if(Gate::denies('policy_kesehatan_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        try {
+            DB::beginTransaction();
 
-        if (count($policyKesehatan->upload_dokumen) > 0) {
-            foreach ($policyKesehatan->upload_dokumen as $media) {
-                if (! in_array($media->file_name, $request->input('upload_dokumen', []))) {
-                    $media->delete();
+            // Insert ke tabel central
+            $policiesCentral = $policyKesehatan->id_policies;
+            $policiesCentral->update($request->centralData());
+                       
+            if (count($policiesCentral->external_polis_doc) > 0) {
+                foreach ($policiesCentral->external_polis_doc as $media) {
+                    if (! in_array($media->file_name, $request->input('external_polis_doc', []))) {
+                        $media->delete();
+                    }
                 }
             }
-        }
-        $media = $policyKesehatan->upload_dokumen->pluck('file_name')->toArray();
-        foreach ($request->input('upload_dokumen', []) as $file) {
-            if (count($media) === 0 || ! in_array($file, $media)) {
-                $policyKesehatan->addMedia(storage_path('tmp/uploads/' . basename($file)))->toMediaCollection('upload_dokumen');
+            $media = $policiesCentral->external_polis_doc->pluck('file_name')->toArray();
+            foreach ($request->input('external_polis_doc', []) as $file) {
+                if (count($media) === 0 || ! in_array($file, $media)) {
+                    $policiesCentral->addMedia(storage_path('tmp/uploads/' . basename($file)))->toMediaCollection('external_polis_doc');
+                }
             }
-        }
 
-        return redirect()->route('admin.policy-kesehatans.index');
+            // Insert ke child sesuai type
+            $childData = $request->childData();
+            $childData['id_policies_id'] = $policiesCentral->id;
+            $policyKesehatan->update($childData);
+
+            if (count($policyKesehatan->upload_dokumen) > 0) {
+                foreach ($policyKesehatan->upload_dokumen as $media) {
+                    if (! in_array($media->file_name, $request->input('upload_dokumen', []))) {
+                        $media->delete();
+                    }
+                }
+            }
+            $media = $policyKesehatan->upload_dokumen->pluck('file_name')->toArray();
+            foreach ($request->input('upload_dokumen', []) as $file) {
+                if (count($media) === 0 || ! in_array($file, $media)) {
+                    $policyKesehatan->addMedia(storage_path('tmp/uploads/' . basename($file)))->toMediaCollection('upload_dokumen');
+                }
+            }
+
+            return redirect()->route('admin.policy-kesehatans.index');
+
+            DB::commit();
+            return redirect()->route('admin.policy-rumah-gedungs.index');
+        } catch (\Throwable $e) {
+            DB::rollBack();
+            \Log::error('Gagal menyimpan polis', ['error' => $e->getMessage()]);
+            return back()->with('error', 'Terjadi kesalahan saat menyimpan data.')->withInput();
+        }
     }
 
     public function show(PolicyKesehatan $policyKesehatan)
